frontend: promote Skills Runner to /skills as new 'Runners' tab
The runner UX was buried at Settings → Developer Options → Skills
Runner. The top-level /skills tab — the discoverable home — had no
way to run anything. Now all 3 bundled skills (github-issue-crusher,
pr-review-shepherd, dev-workflow) are reachable from /skills with
their full picker + Run + Schedule + Recent Runs + log viewer UX.

Three small changes, one shared component:

(1) Extract: SkillsRunnerPanel's body (everything except the Settings
    shell — picker, dynamic input form, Run Now, Schedule cron, Recent
    Runs viewer with click-to-expand log tail) moves into
    app/src/components/skills/SkillsRunnerBody.tsx as a reusable
    component. Renamed the descriptive-header prop to `headerText`
    to avoid shadowing the internal `description` state that holds
    the resolved SkillDescription.

(2) Slim: settings/panels/SkillsRunnerPanel.tsx becomes a 30-line
    thin wrapper around <SkillsRunnerBody /> — keeps the existing
    /settings/skills-runner route working as a shortcut.

(3) Promote: pages/Skills.tsx PillTabBar gets a new 'Runners' tab.
    Renders <SkillsRunnerBody /> in a card alongside the existing
    Composio / Channels / MCP tabs. Bottom of the card has a small
    blurb linking to /settings/dev-workflow for the specialized
    cron-driven dev-workflow setup (its repo / fork / branch picker
    doesn't generalize; left in place rather than ported wholesale).

3 new i18n keys: skills.tabs.runners + skills.runners.specialized.*.
Locale-chunk parity still deferred (pnpm i18n:check not wired on
this branch).

After this commit /skills is the canonical home for skills work:
browse / install / create the catalog (existing), pick + run +
schedule + view history of bundled runners (new).
