+++ app/src/lib/i18n/en.ts
@@ -260,6 +260,10 @@ const en: TranslationMap = {
   'skills.tabs.composio': 'Composio',
   'skills.tabs.channels': 'Channels',
   'skills.tabs.mcp': 'MCP Servers',
+  'skills.tabs.runners': 'Runners',
+  'skills.runners.specialized.devWorkflowBlurb':
+    'Looking for the recurring autonomous-developer workflow with a repo / fork / branch picker?',
+  'skills.runners.specialized.openDevWorkflow': 'Open Dev Workflow setup →',
 
   // Intelligence / Memory
   'memory.title': 'Memory',
